pyproject.toml
[project]
name = "buff"
version = "0.1.0"
description = "Deterministic OHLCV data pipeline: Binance Futures candles to parquet + quality reports"
requires-python = ">=3.10"
dependencies = [
    "ccxt>=1.91.0",
    "pandas>=1.5.0",
    "pyarrow>=10.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Fast unit tests that don't require I/O or network",
    "integration: Slower integration tests with fixtures like tmp_path",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"
